requirements.txt
spoon-ai-sdk>=0.3.0
spoon-toolkits>=0.2.0
PyPDF2>=3.0.0
requests>=2.31.0
anthropic>=0.34.0
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.0.0
werkzeug>=3.0.0

